Prepare Physical Memory
REMOVED - All of the lab 3 testing 
MODIFIED - Main now goes through and builds the kernel page table and sort of adds it to physical memory
NEED - figure out what to pass the memory allocator in terms of the vector that used to refer to processes. See my todo comment. If you could figure that out it would be really helpful